+++ Demo/JSONPreviewDemo/Core/Global/CustomStyleExampleViewController.swift
@@ -10,7 +10,7 @@ import UIKit
 
 import JSONPreview
 
-class CustomStyleExampleViewController: BaseJSONPreviewController {
+final class CustomStyleExampleViewController: BaseJSONPreviewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         
@@ -32,7 +32,7 @@ private extension CustomStyleExampleViewController {
         
         NSLayoutConstraint.activate([
             previewView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor),
-            previewView.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor),
+            previewView.bottomAnchor.constraint(equalTo: view.bottomAnchor),
             previewView.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor),
             previewView.trailingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.trailingAnchor)
         ])