+++ Demo/JSONPreviewDemo/Core/Global/EntranceTableViewController.swift
@@ -23,8 +23,28 @@ final class EntranceTableViewController: ListTableViewController {
             ),
         ]
         
-#if !os(tvOS)
+        let tableViewExample = DemoCaseConfig(
+            title: "In UITableView",
+            desc: "Show how to use it in UITableView and dynamically adapt to Cell height.",
+            action: { TableViewExampleViewController() }
+        )
+        
+#if os(tvOS)
         dataSource.append(contentsOf: [
+            .init(
+                title: "Wrap",
+                desc: "This example shows the rendering and interactive effects of JSONPreview when displaying JSON without breaking lines. This example is built specifically for tvOS.",
+                action: { TVOSWrapExampleViewController() }
+            ),
+            tableViewExample,
+        ])
+#else
+        dataSource.append(contentsOf: [
+            .init(
+                title: "Automatic Wrap",
+                desc: "This example mainly demonstrates the rendering effect of JSON when lines are automatically wrapped and not.",
+                action: { WrapExampleViewController() }
+            ),
             .init(
                 title: "Initial folded",
                 desc: "Initial collapse of all JSON nodes. You need to expand each node manually.",
@@ -35,19 +55,25 @@ final class EntranceTableViewController: ListTableViewController {
                 desc: "Example of how to hide line numbers.",
                 action: { HideLineNumberExampleViewController() }
             ),
-            .init(
-                title: "In UITableView",
-                desc: "Show how to use it in UITableView and dynamically adapt to Cell height.",
-                action: { TableViewExampleViewController() }
-            ),
+            tableViewExample,
             .init(
                 title: "Seach",
                 desc: "Related examples of search functionality.",
                 action: { SearchEntranceTableViewController() }
-            )
+            ),
         ])
 #endif
-        super.init(dataSource: dataSource)
+        
+        super.init(dataSource: [
+            dataSource,
+            [
+                .init(
+                    title: "Test",
+                    desc: "Not used to demonstrate functionality, but used for internal testing.",
+                    action: { TestEntranceTableViewController() }
+                )
+            ]
+        ])
     }
     
     required init?(coder: NSCoder) {